// apps/api/src/generated/prisma/enums.ts
/* !!! This is code generated by Prisma. Do not edit directly. !!! */
/* eslint-disable */
// biome-ignore-all lint: generated file
// @ts-nocheck 
/*
* This file exports all enum related types from the schema.
*
* 🟢 You can import this file directly.
*/

export const ProviderKind = {
  LLM: 'LLM',
  EMBEDDING: 'EMBEDDING'
} as const

export type ProviderKind = (typeof ProviderKind)[keyof typeof ProviderKind]


export const SkillType = {
  BUILTIN_FUNCTION: 'BUILTIN_FUNCTION',
  HTTP_RPA: 'HTTP_RPA',
  EXTERNAL_AGENT: 'EXTERNAL_AGENT'
} as const

export type SkillType = (typeof SkillType)[keyof typeof SkillType]


export const TemplateStatus = {
  DRAFT: 'DRAFT',
  PENDING_REVIEW: 'PENDING_REVIEW',
  PUBLISHED: 'PUBLISHED',
  ARCHIVED: 'ARCHIVED',
  REJECTED: 'REJECTED'
} as const

export type TemplateStatus = (typeof TemplateStatus)[keyof typeof TemplateStatus]


export const KbMode = {
  INJECT: 'INJECT',
  TOOL: 'TOOL',
  BOTH: 'BOTH'
} as const

export type KbMode = (typeof KbMode)[keyof typeof KbMode]


export const DocSourceType = {
  UPLOAD: 'UPLOAD',
  CRAWL: 'CRAWL',
  MANUAL: 'MANUAL'
} as const

export type DocSourceType = (typeof DocSourceType)[keyof typeof DocSourceType]


export const DocStatus = {
  UPLOADED: 'UPLOADED',
  QUEUED: 'QUEUED',
  PARSING: 'PARSING',
  INDEXED: 'INDEXED',
  FAILED: 'FAILED',
  OCR_NEEDED: 'OCR_NEEDED'
} as const

export type DocStatus = (typeof DocStatus)[keyof typeof DocStatus]


export const MemberRole = {
  OWNER: 'OWNER',
  ADMIN: 'ADMIN',
  MEMBER: 'MEMBER'
} as const

export type MemberRole = (typeof MemberRole)[keyof typeof MemberRole]


export const GrantScopeType = {
  ENTERPRISE: 'ENTERPRISE',
  DEPARTMENT: 'DEPARTMENT',
  MEMBER: 'MEMBER'
} as const

export type GrantScopeType = (typeof GrantScopeType)[keyof typeof GrantScopeType]


export const MessageRole = {
  USER: 'USER',
  ASSISTANT: 'ASSISTANT',
  SYSTEM: 'SYSTEM',
  TOOL: 'TOOL'
} as const

export type MessageRole = (typeof MessageRole)[keyof typeof MessageRole]


export const UsageKind = {
  LLM: 'LLM',
  EMBEDDING: 'EMBEDDING',
  RPA: 'RPA',
  EXTERNAL_AGENT: 'EXTERNAL_AGENT'
} as const

export type UsageKind = (typeof UsageKind)[keyof typeof UsageKind]